Move populator to separate package and start populator in controller
